ChronoSim

Documentation for ChronoSim.

ChronoSim.PhysicalState

ChronoSim.capture_state_changes

ChronoSim.capture_state_reads

ChronoSim.isconsistent

ChronoSim.over_generated_events

ChronoSim.over_tracked_physical_state

ChronoSim.PhysicalState — Type

PhysicalState is an abstract type from which to inherit the state of a simulation. A PhysicalState should put all mutable values, the values upon which events depend, into TrackedVector objects. For instance:

@tracked_struct Square begin
    occupant::Int
    resistance::Float64
end

# Everything we know about an agent.
@tracked_struct Agent begin
    health::Health
    loc::CartesianIndex{2}
end

mutable struct BoardState <: PhysicalState
    board::TrackedVector{Square}
    agent::TrackedVector{Agent}
end

The PhysicalState may contain other properties, but those defined with TrackedVectors are used to compute the next event in the simulation.

ChronoSim.capture_state_changes — Method

capture_state_changes(f::Function, physical_state)

The callback function f will modify the physical state. This function records which parts of the state were modified. The callback should have no arguments and may return a result.

ChronoSim.capture_state_reads — Method

capture_state_reads(f::Function, physical_state)

The callback function f will read the physical state. This function records which parts of the state were read. The callback should have no arguments and may return a result.

ChronoSim.isconsistent — Method

isconsistent(physical_state)

A simulation in debug mode will assert isconsistent(physical_state) is true. Override this to verify the physical state of your simulation.

ChronoSim.over_generated_events — Method

over_generated_events(f::Function, generators, physical, fired_event_key, changed_places)

Given a fired event and the set of places changed by that event, create new events that may depend on that fired event and those changed places.

ChronoSim.over_tracked_physical_state — Method

over_tracked_physical_state(fcallback::Function, physical::PhysicalState)

Iterate over all tracked vectors in the physical state.