Reference

ChronoSim.DepNetNaive

ChronoSim.DependencyNetwork

ChronoSim.EventGenerator

ChronoSim.InitializeEvent

ChronoSim.Member

ChronoSim.PhysicalState

ChronoSim.SimEvent

ChronoSim.SimulationFSM

ChronoSim.ObservedState.access_to_placekey

ChronoSim.access_to_argnames

ChronoSim.access_to_searchkey

ChronoSim.capture_state_changes

ChronoSim.capture_state_changes

ChronoSim.capture_state_reads

ChronoSim.capture_state_reads

ChronoSim.clock_key

ChronoSim.common_base_key_tuple

ChronoSim.deal_with_changes

ChronoSim.escaped_args

ChronoSim.fire!

ChronoSim.generators

ChronoSim.initialize!

ChronoSim.isconsistent

ChronoSim.isimmediate

ChronoSim.key_clock

ChronoSim.over_generated_events

ChronoSim.over_tracked_physical_state

ChronoSim.run

ChronoSim.trace_likelihood

ChronoSim.@conditionsfor

ChronoSim.@reactto

ChronoSim.ObservedState.@keyedby

ChronoSim.ObservedState.@observe

ChronoSim.ObservedState.@observedphysical

ChronoSim.DepNetNaive — Type

For testing, we make an equivalent version of the dependency network but this one uses a very different internal structure, an edge list.

ChronoSim.DependencyNetwork — Type

This is a multi-graph from places to events. It is a multi-graph because there can be two kinds of edges from the same place to the same event: an enabling edge and a rate edge. The graph is mutable, too, so you can add and remove events which adds and removes edges from the graph.

ChronoSim.EventGenerator — Type

EventGenerator(match_what, matchstr, generator::Function)

When an event fires, it changes the physical state. The simulation observes which parts of the physical state changed and sends those parts to this EventGenerator. The EventGenerator is a rule that matches changes to the physical state and creates SimEvent that act on that physical state.

The matchstr is a list of symbols (array_name, ℤ, struct_member). The ℤ represents the integer index within the array. For instance, if we simulated chess, it might be (:board, ℤ, :piece).

The generator is a callback that the simulation uses to determine which events need to be enabled given recent changes to the state of the board. Its signature is:

    callback_function(f::Function, physical_state, indices...)

Here the indices are the integer index that matches the ℤ above. This callback function should look at the physical state and call f(transition) where transition is an instance of SimEvent.

ChronoSim.InitializeEvent — Type

InitializeEvent is a concrete transition type that represents the first event in the system, initialization.

ChronoSim.Member — Type

This represents a field in a struct. It is a wrapper around Symbol. We wrap the Symbol so that it doesn't conflict with dictionary keys that are symbols.

ChronoSim.PhysicalState — Type

PhysicalState is an abstract type from which to inherit the state of a simulation. A PhysicalState should put all mutable values, the values upon which events depend, into TrackedVector objects. For instance:

@tracked_struct Square begin
    occupant::Int
    resistance::Float64
end

# Everything we know about an agent.
@tracked_struct Agent begin
    health::Health
    loc::CartesianIndex{2}
end

mutable struct BoardState <: PhysicalState
    board::TrackedVector{Square}
    agent::TrackedVector{Agent}
end

The PhysicalState may contain other properties, but those defined with TrackedVectors are used to compute the next event in the simulation.

ChronoSim.SimEvent — Type

SimEvent

This abstract type is the parent of all transitions in the system.

ChronoSim.SimulationFSM — Method

SimulationFSM(physical_state, trans_rules; seed, rng, sampler, observer=nothing)

Create a simulation.

The physical_state is of type PhysicalState. The sampler is of type CompetingClocks.SSA. The trans_rules are a list of type SimEvent. The seed is an integer seed for a Xoshiro random number generator. The observer is a callback with the signature:

observer(physical, when::Float64, event::SimEvent, changed_places::AbstractSet{Tuple})

The changed_places argument is a set-like object with tuples that are keys that represent which places were changed.

ChronoSim.access_to_argnames — Method

access_to_argnames(expr)

It extracts variable arguments to an accessor expression. An accessor expression uses getfield or getindex to find members of structs and containers. Turns this: :(agent[j].value[k]), into this: [:(j), :(k)].

ChronoSim.access_to_searchkey — Method

access_to_searchkey(expr::Expr)

Convert a Julia access expression into a search key pattern for the event generator system. Field names become Member objects, and array/container indices become MEMBERINDEX placeholders.

Examples

# Simple member access
julia> access_to_searchkey(:(obj.field))
[Member(:obj), Member(:field)]

# Array access with index
julia> access_to_searchkey(:(arr[5].field))  
[Member(:arr), MEMBERINDEX, Member(:field)]

# Multi-dimensional array access
julia> access_to_searchkey(:(board[i][j].piece))
[Member(:board), MEMBERINDEX, MEMBERINDEX, Member(:piece)]

The resulting pattern is used to match against place keys in the simulation's state change tracking.

ChronoSim.capture_state_changes — Method

capture_state_changes(f::Function, physical_state)

The callback function f will modify the physical state. This function records which parts of the state were modified. The callback should have no arguments and may return a result.

ChronoSim.capture_state_reads — Method

capture_state_reads(f::Function, physical_state)

The callback function f will read the physical state. This function records which parts of the state were read. The callback should have no arguments and may return a result.

ChronoSim.clock_key — Method

clock_key(::SimEvent)::Tuple

All SimEvent objects are immutable structs that represent events but don't carry any mutable state. A clock key is a tuple version of an event.

ChronoSim.common_base_key_tuple — Method

Look at events and determine a common base type. Internally the simulation tracks events with sets of tuples by turning each event instance into a tuple. If all the tuples have the same type, this should turn out to be performant.

ChronoSim.deal_with_changes — Method

deal_with_changes(sim::SimulationFSM)

An event changed the state. This function modifies events to respond to changes in state.

ChronoSim.escaped_args — Method

Instead of escaping the argument list as-is, we reach into arguments that are tuples and escape them individually. The idea is to destructure the arguments into escaped variables.

ChronoSim.fire! — Method

fire!(sim::SimulationFSM, time, event_key)

Let the event act on the state.

ChronoSim.generators — Method

generators(::Type{SimEvent})::Vector{EventGenerator}

Every transition in the simulation needs generators that notice changes to state or events fired and create the appropriate transitions. Implement a generators function as part of the interface of each transition.

ChronoSim.initialize! — Method

Initialize the simulation. You could call it as a do-function. It is structured this way so that the simulation will record changes to the physical state.

    initialize!(sim) do init_physical
        initialize!(init_physical, agent_cnt, sim.rng)
    end

ChronoSim.isconsistent — Method

isconsistent(physical_state)

A simulation in debug mode will assert isconsistent(physical_state) is true. Override this to verify the physical state of your simulation.

ChronoSim.isimmediate — Method

isimmediate(EventType)

An immediate event should return true for this function.

ChronoSim.key_clock — Method

key_clock(key::Tuple, event_dict::Dict{Symbol, DataType})::SimEvent

Takes a tuple of the form (:symbol, arg, arg) and a dictionary mapping symbols to struct types, and returns an instantiation of the struct named by :symbol. We pass in the list of datatypes because, if we didn't, then instantiation of a type from a symbol would need to search for the correct constructor in the correct module, and that would be both wrong and slow.

ChronoSim.over_generated_events — Method

over_generated_events(f::Function, generators, physical, fired_event_key, changed_places)

Given a fired event and the set of places changed by that event, create new events that may depend on that fired event and those changed places.

ChronoSim.over_tracked_physical_state — Method

over_tracked_physical_state(fcallback::Function, physical::PhysicalState)

Iterate over all tracked vectors in the physical state.

ChronoSim.run — Method

run(simulation, initializer, stop_condition)

Given a simulation, this initializes the physical state and generates a trajectory from the simulation until the stop condition is met. The initializer is either a function whose argument is a physical state and returns nothing, or it is an event key for an event that initializes the system. The stop condition is a function with the signature:

stop_condition(physical_state, step_idx, event::SimEvent, when)::Bool

The event and when passed into the stop condition are the event and time that are about to fire but have not yet fired. This lets you enforce a stopping time that is between events.

ChronoSim.trace_likelihood — Method

The trace is a Vector{Tuple{Float64,SimEvent}}. That is, it's a list of tuples containing (when, what event). In order to calculate log-likelihood of a simulation, pass it a sampler that tracks log-likelihood. For instance,

base_sampler = CombinedNextReaction{K,T}()
memory_sampler = MemorySampler(base_sampler)

ChronoSim.@conditionsfor — Macro

@conditionsfor EventType begin
    @reactto ... end
    @reactto ... end
end

Generates a generators(::Type{EventType}) function containing all the EventGenerators defined in the @reactto blocks.

ChronoSim.@reactto — Macro

@reactto changed(array[index].field) do physical
    # generator body
end

@reactto fired(EventType(args...)) do physical
    # generator body
end

Creates an EventGenerator that reacts to state changes or event firings. Used within @conditionsfor blocks.

ChronoSim.ObservedState.access_to_placekey — Method

access_to_placekey(expr)

This function takes an expression for member access to a hierarchical container and converts it into a place key. For instance:

Accessor expression	Place key

state.agent[j].armor	(:agent, j, :armor)

sim.board[i, j].fval	(:board, (i, j), :fval)

state.param	(:param,)

sim.adict[(name, kind)]	(:adict, (name, kind))

physical.land[square].grass	(:land, square, :grass)

The output is also an Expr-type object.

ChronoSim.capture_state_changes — Method

capture_state_changes(f::Function, physical_state)

The callback function f will modify the physical state. This function records which parts of the state were modified. The callback should have no arguments and may return a result.

ChronoSim.capture_state_reads — Method

capture_state_reads(f::Function, physical_state)

The callback function f will read the physical state. This function records which parts of the state were read. The callback should have no arguments and may return a result.

ChronoSim.ObservedState.@keyedby — Macro

@keyedby StructName IndexType begin
    field1::Type1
    field2::Type2
    ...
end

Create a mutable struct with the given fields plus automatic _container::Any and _index::IndexType fields. The generated constructor only requires the user-defined fields.

Example

@keyedby MyElement Int64 begin
    val::Int64
    name::String
end

This generates a struct equivalent to:

mutable struct MyElement
    val::Int64
    name::String
    _container::Any
    _index::Int64
    MyElement(val, name) = new(val, name)
end

ChronoSim.ObservedState.@observe — Macro

@observe expr

Track reads and writes to ObservedPhysical state. For reads, records the access in obsread and returns the value. For writes, records the access in obsmodified and performs the assignment.

ChronoSim.ObservedState.@observedphysical — Macro

@observedphysical <typename> <definition block>

This macro defines a physical state that uses the ObservedState machinery to track changes to that physical state.

For example:

@keyedby Int Piece begin
    speed::Float64
    kind::String
end

@keyedby NTuple{2,Int64} Square begin
    grass::Float64
    food::Float64
end

@observedphysical Board begin
    board::ObservedArray{Square,2}
    actor::ObservedDict{Int,Piece}
    params::Dict{Symbol,Float64}
    actors_max::Int64
end

This macro creates a struct that contains tracking information about its Observed members, so in the example params would not be tracked. The Board type has a parent type ObservedPhysical which has the parent type ChronoSim.PhysicalState.

Expanding the macro, we would see:

mutable struct Board <: ObservedPhysical
    board::ObservedArray{Square,2}
    actor::ObservedDict{Int,Piece}
    params::Dict{Symbol,Float64}
    actors_max::Int64
    obs_modified::Vector{Tuple}
    obs_read::Vector{Tuple}
    Board(board, actor, params, actors_max) = ...
end

Here, the list of read and modified state uses the abstract type Tuple because it unifies the way changed state is reported for the two observed containers: (:board, (2,2), :grass) and (:actor, 7, :speed). The ellipsis above for the constructor is a constructor that initializes obs_modified and obs_read to empty vectors.